+++ test/infrastructure/builders/resource_type_builder.go
@@ -42,6 +42,37 @@ func (b *ResourceTypeBuilder) WithResponseBodyIDFilter() *ResourceTypeBuilder {
 	return b
 }
 
+func (b *ResourceTypeBuilder) WithQueryParameters() *ResourceTypeBuilder {
+	b.attributeTypes["query_parameters"] = tftypes.Map{ElementType: tftypes.String}
+	return b
+
+}
+
+func (b *ResourceTypeBuilder) WithID() *ResourceTypeBuilder {
+	b.attributeTypes["id"] = tftypes.String
+	return b
+}
+
+func (b *ResourceTypeBuilder) WithResponseCode() *ResourceTypeBuilder {
+	b.attributeTypes["response_code"] = tftypes.Number
+	return b
+}
+
+func (b *ResourceTypeBuilder) WithResponseBody() *ResourceTypeBuilder {
+	b.attributeTypes["response_body"] = tftypes.String
+	return b
+}
+
+func (b *ResourceTypeBuilder) WithResponseBodyID() *ResourceTypeBuilder {
+	b.attributeTypes["response_body_id"] = tftypes.String
+	return b
+}
+
+func (b *ResourceTypeBuilder) WithResponseBodyJSON() *ResourceTypeBuilder {
+	b.attributeTypes["response_body_json"] = tftypes.Map{ElementType: tftypes.String}
+	return b
+}
+
 func (b *ResourceTypeBuilder) Build() tftypes.Object {
 	return tftypes.Object{
 		AttributeTypes: b.attributeTypes,